+++ plugins/rmanipulation/taskmanipulation.cpp
@@ -617,6 +617,8 @@ protected:
             _robot->SetActiveDOFs(pmanip->GetGripperIndices(), DOF_NoTransform);
             _robot->SetActiveDOFValues(vgoalpreshape,true);
 
+            dReal fGraspApproachOffset = fApproachOffset;
+
             if( !!_pGrasperPlanner && pmanip->GetIkSolver()->Supports(IKP_Transform6D) ) {
                 _robot->SetActiveDOFs(pmanip->GetGripperIndices(), DOF_X|DOF_Y|DOF_Z);
                 if( !_phandtraj ) {
@@ -628,11 +630,11 @@ protected:
                 graspparams->ftargetroll = pgrasp[iGraspRoll];
                 graspparams->vtargetdirection = Vector(pgrasp[iGraspDir], pgrasp[iGraspDir+1], pgrasp[iGraspDir+2]);
                 graspparams->vtargetposition = Vector(pgrasp[iGraspPos], pgrasp[iGraspPos+1], pgrasp[iGraspPos+2]);
-                if( imanipulatordirection < 0 ) {
-                    graspparams->vmanipulatordirection = pmanip->GetDirection();
+                if( imanipulatordirection >= 0 ) {
+                    graspparams->vmanipulatordirection = Vector(pgrasp[imanipulatordirection], pgrasp[imanipulatordirection+1], pgrasp[imanipulatordirection+2]);
                 }
                 else {
-                    graspparams->vmanipulatordirection = Vector(pgrasp[imanipulatordirection], pgrasp[imanipulatordirection+1], pgrasp[imanipulatordirection+2]);
+                    graspparams->vmanipulatordirection = pmanip->GetDirection();
                 }
                 graspparams->btransformrobot = true;
                 graspparams->breturntrajectory = false;
@@ -667,10 +669,16 @@ protected:
                     else {
                         vglobalpalmdir = pmanip->GetTransform().rotate(pmanip->GetDirection());
                     }
-
+                    dReal fstep=0;
+                    dReal fstepbacksize = 0.001f;
                     while(GetEnv()->CheckCollision(KinBodyConstPtr(_robot),KinBodyConstPtr(ptarget))) {
-                        t.trans -= vglobalpalmdir*0.001f;
+                        t.trans -= vglobalpalmdir*fstepbacksize;
+                        fGraspApproachOffset -= fstepbacksize;
+                        fstep += fstepbacksize;
                         _robot->SetTransform(t);
+                    }
+                    if( fstep > 0 ) {
+                        RAVELOG_DEBUG(str(boost::format("grasp %d: moved %f along direction=[%f,%f,%f]")%igrasp%fstep% -vglobalpalmdir.x% -vglobalpalmdir.y% -vglobalpalmdir.z));
                     }
                 }
 
@@ -700,7 +708,7 @@ protected:
                     tGoalEndEffector.SetTranslationDirection5D(RAY(tgoal.trans,tgoal.rotate(pmanip->GetDirection())));
                     vector<dReal> vsolution;
                     if( !pmanip->FindIKSolution(tGoalEndEffector,vsolution,IKFO_CheckEnvCollisions) ) {
-                        RAVELOG_DEBUG("ik 5d failed: %d\n", igrasp);
+                        RAVELOG_DEBUG(str(boost::format("grasp %d: ik 5d failed")%igrasp));
                         continue; // failed
                     }
                     vFinalGripperValues = _vFinalGripperValues;
@@ -719,7 +727,7 @@ protected:
                 }
             }
             else {
-                RAVELOG_ERROR("grasper problem not valid\n");
+                RAVELOG_ERROR(str(boost::format("grasp %d: grasper problem not valid")%igrasp));
                 return false;
             }
 
@@ -745,12 +753,6 @@ protected:
                     }
                 }
 
-                dReal fSmallOffset = 0.0001f;
-                if( fApproachOffset != 0 ) {
-                    Transform tsmalloffset; tsmalloffset.trans = -fSmallOffset * vglobalpalmdir;
-                    tApproachEndEffector = tsmalloffset*tApproachEndEffector;
-                }
-
                 // first test the IK solution at the destination tGoalEndEffector
                 if( !pmanip->FindIKSolution(tApproachEndEffector, viksolution, IKFO_CheckEnvCollisions) ) {
                     RAVELOG_DEBUG("grasp %d: No IK solution found (final)\n", igrasp);
@@ -758,10 +760,10 @@ protected:
                 }
 
                 _UpdateSwitchModels(true,true);     // switch to fat models
-                if( fApproachOffset != 0 ) {
+                if( fGraspApproachOffset > 0 ) {
                     Transform tsmalloffset;
                     // now test at the approach point (with offset)
-                    tsmalloffset.trans = -(fApproachOffset-fSmallOffset) * vglobalpalmdir;
+                    tsmalloffset.trans = -fGraspApproachOffset * vglobalpalmdir;
                     tApproachEndEffector = tsmalloffset*tApproachEndEffector;
 
                     // set the previous robot ik configuration to get the closest configuration!!
@@ -773,12 +775,12 @@ protected:
                         continue;
                     }
                     else {
-                        stringstream ss;
-                        ss << "IK found: ";
+                        stringstream ss; ss << std::setprecision(std::numeric_limits<dReal>::digits10+1);
+                        ss << "ikfound = [";
                         FOREACH(it, viksolution) {
-                            ss << *it << " ";
+                            ss << *it << ", ";
                         }
-                        ss << endl;
+                        ss << "]" << endl;
                         RAVELOG_DEBUG(ss.str());
                     }
                 }